Put all the manager controller together
