+++ src/main/resources/templates/jobs/manager/completed_jobs.html
