+++ src/main/resources/templates/jobs/manager/index.html
