+++ src/main/resources/templates/jobs/manager/job_add.html
