+++ src/main/resources/templates/jobs/manager/job_edit_manager.html
