+++ src/main/resources/templates/lots/manager/index.html
