+++ src/main/resources/templates/operation/manager/create_operation.html
