+++ src/main/resources/templates/operation/manager/edit.html
