+++ src/main/resources/templates/operation/manager/index.html
