+++ src/main/resources/templates/parts/manager/index.html
