+++ src/main/resources/templates/product/manager/create_product.html
