+++ src/main/resources/templates/product/manager/edit.html
