+++ src/main/resources/templates/product/manager/index.html
